move BuildPhase
