+++ Sources/ProjectSpec/TargetSource.swift
@@ -1,8 +1,6 @@
 import Foundation
 import JSONUtilities
 import PathKit
-import enum XcodeProj.BuildPhase
-import class XcodeProj.PBXCopyFilesBuildPhase
 
 public struct TargetSource: Equatable {
     public static let optionalDefault = false
@@ -15,7 +13,7 @@ public struct TargetSource: Equatable {
     public var includes: [String]
     public var type: SourceType?
     public var optional: Bool
-    public var buildPhase: BuildPhase?
+    public var buildPhase: BuildPhaseSpec?
     public var headerVisibility: HeaderVisibility?
     public var createIntermediateGroups: Bool?
     public var attributes: [String]
@@ -35,94 +33,6 @@ public struct TargetSource: Equatable {
         }
     }
 
-    public enum BuildPhase: Equatable {
-        case sources
-        case headers
-        case resources
-        case copyFiles(CopyFilesSettings)
-        case none
-        // Not currently exposed as selectable options, but used internally
-        case frameworks
-        case runScript
-        case carbonResources
-
-        public struct CopyFilesSettings: Equatable, Hashable {
-            public static let xpcServices = CopyFilesSettings(
-                destination: .productsDirectory,
-                subpath: "$(CONTENTS_FOLDER_PATH)/XPCServices",
-                phaseOrder: .postCompile
-            )
-
-            public enum Destination: String {
-                case absolutePath
-                case productsDirectory
-                case wrapper
-                case executables
-                case resources
-                case javaResources
-                case frameworks
-                case sharedFrameworks
-                case sharedSupport
-                case plugins
-
-                public var destination: PBXCopyFilesBuildPhase.SubFolder? {
-                    switch self {
-                    case .absolutePath: return .absolutePath
-                    case .productsDirectory: return .productsDirectory
-                    case .wrapper: return .wrapper
-                    case .executables: return .executables
-                    case .resources: return .resources
-                    case .javaResources: return .javaResources
-                    case .frameworks: return .frameworks
-                    case .sharedFrameworks: return .sharedFrameworks
-                    case .sharedSupport: return .sharedSupport
-                    case .plugins: return .plugins
-                    }
-                }
-            }
-
-            public enum PhaseOrder: String {
-                /// Run before the Compile Sources phase
-                case preCompile
-                /// Run after the Compile Sources and post-compile Run Script phases
-                case postCompile
-            }
-
-            public var destination: Destination
-            public var subpath: String
-            public var phaseOrder: PhaseOrder
-
-            public init(
-                destination: Destination,
-                subpath: String,
-                phaseOrder: PhaseOrder
-            ) {
-                self.destination = destination
-                self.subpath = subpath
-                self.phaseOrder = phaseOrder
-            }
-        }
-
-        public var buildPhase: XcodeProj.BuildPhase? {
-            switch self {
-            case .sources: return .sources
-            case .headers: return .headers
-            case .resources: return .resources
-            case .copyFiles: return .copyFiles
-            case .frameworks: return .frameworks
-            case .runScript: return .runScript
-            case .carbonResources: return .carbonResources
-            case .none: return nil
-            }
-        }
-    }
-
-    public enum SourceType: String {
-        case group
-        case file
-        case folder
-    }
-
     public init(
         path: String,
         name: String? = nil,
@@ -132,7 +42,7 @@ public struct TargetSource: Equatable {
         includes: [String] = [],
         type: SourceType? = nil,
         optional: Bool = optionalDefault,
-        buildPhase: BuildPhase? = nil,
+        buildPhase: BuildPhaseSpec? = nil,
         headerVisibility: HeaderVisibility? = nil,
         createIntermediateGroups: Bool? = nil,
         attributes: [String] = [],
@@ -188,9 +98,9 @@ extension TargetSource: JSONObjectConvertible {
         optional = jsonDictionary.json(atKeyPath: "optional") ?? TargetSource.optionalDefault
 
         if let string: String = jsonDictionary.json(atKeyPath: "buildPhase") {
-            buildPhase = try BuildPhase(string: string)
+            buildPhase = try BuildPhaseSpec(string: string)
         } else if let dict: JSONDictionary = jsonDictionary.json(atKeyPath: "buildPhase") {
-            buildPhase = try BuildPhase(jsonDictionary: dict)
+            buildPhase = try BuildPhaseSpec(jsonDictionary: dict)
         }
 
         createIntermediateGroups = jsonDictionary.json(atKeyPath: "createIntermediateGroups")
@@ -225,62 +135,6 @@ extension TargetSource: JSONEncodable {
         dict["path"] = path
 
         return dict
-    }
-}
-
-extension TargetSource.BuildPhase {
-
-    public init(string: String) throws {
-        switch string {
-        case "sources": self = .sources
-        case "headers": self = .headers
-        case "resources": self = .resources
-        case "copyFiles":
-            throw SpecParsingError.invalidSourceBuildPhase("copyFiles must specify a \"destination\" and optional \"subpath\"")
-        case "none": self = .none
-        default:
-            throw SpecParsingError.invalidSourceBuildPhase(string.quoted)
-        }
-    }
-}
-
-extension TargetSource.BuildPhase: JSONObjectConvertible {
-
-    public init(jsonDictionary: JSONDictionary) throws {
-        self = .copyFiles(try jsonDictionary.json(atKeyPath: "copyFiles"))
-    }
-}
-
-extension TargetSource.BuildPhase: JSONEncodable {
-    public func toJSONValue() -> Any {
-        switch self {
-        case .sources: return "sources"
-        case .headers: return "headers"
-        case .resources: return "resources"
-        case .copyFiles(let files): return ["copyFiles": files.toJSONValue()]
-        case .none: return "none"
-        case .frameworks: fatalError("invalid build phase")
-        case .runScript: fatalError("invalid build phase")
-        case .carbonResources: fatalError("invalid build phase")
-        }
-    }
-}
-
-extension TargetSource.BuildPhase.CopyFilesSettings: JSONObjectConvertible {
-
-    public init(jsonDictionary: JSONDictionary) throws {
-        destination = try jsonDictionary.json(atKeyPath: "destination")
-        subpath = jsonDictionary.json(atKeyPath: "subpath") ?? ""
-        phaseOrder = .postCompile
-    }
-}
-
-extension TargetSource.BuildPhase.CopyFilesSettings: JSONEncodable {
-    public func toJSONValue() -> Any {
-        [
-            "destination": destination.rawValue,
-            "subpath": subpath,
-        ]
     }
 }
